# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invillia Test",
        "description": "Invillia Test",
        "contact": {
            "email": "jonathanbeltrao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register New User",
                "description": "Create new User",
                "operationId": "registerUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jonathan",
                                    "email": "jonathan@gmail.com",
                                    "password": "jSuds(*#@*"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Authorization Token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Type of Token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNWY2YmI2N2QzZDRiNTNkODFkZDZkN2ZiOTk1YmZhNzBmMzgwODNmMDk4MGVjYzI2MTQ4MjhiNzQ1ODcyY2UxOTM3NDI3NThhZDVlNzIwMmQiLCJpYXQiOjE2MDc0NDYxODUsIm5iZiI6MTYwNzQ0NjE4NSwiZXhwIjoxNjM4OTgyMTg1LCJzdWIiOiIyIiwic2NvcGVzIjpbXX0.fcnYNE8v8KQX_K_1WZZB6MWWTS3IASu1CV8RmyBXKc4i0aAw_3OfU_91InzLw3PJrPRJm-pmSPFGb19E9J_QyOGOPBwrQoaUIF6Ir3PTFBEpT8P9geswFz3w53HQcqF954MgwM9-JI7aZYsqHRdJu1blUd8vquuhPsLVT0EKvKT7kIZir1IutuArQq1-xTtb_myh7I_r0-s1GhFEh-wBYU7R5GHFyAT_qmYDns9M_J-suLWMeXcugBDiYOBTrWFv_54TwJIW8L-RYrRXdetekn0ndxXqn4StPvAUpXTBnpqTphPHF2o8lIy0wptCgzrQsV5VXzkj80pbWPZDY01O-I1aMpHSCDwzU4BH3It2LLxLezsln881NkPRD5d-hXQeDCJ88ZnTGk5H3pcUxGrhHs8dYFU9CQ9C1EeKuVYt3XsT1Cbo9Ha-qN-xly_4WbIbNnltnE_AOo92PvWeOXrvSoFCbi6yTKR_ygJskFDW5menTIzEbKbmcvwhfdDPU6iN_V7nA2uQXGt6ApKzj5WsXwUiTaZpPkivEc7n1dF_nriz93AsaVtrB_dmuEsS28q8zIMt3F9uYSwwFS0jL3vWvGhxUBmWy440I6RePWEnTc_fTW-o_BC-iBvhNspwGyR4QySAnP48aMPD0muHij_Q-ZEMafKR_26kFSUC08UXFOA",
                                        "token_type": "Bearer"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "description": "Login user to Generate Token",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jonathan",
                                    "email": "jonathan@gmail.com",
                                    "password": "jSuds(*#@*"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "Authorization Token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Type of Token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNWY2YmI2N2QzZDRiNTNkODFkZDZkN2ZiOTk1YmZhNzBmMzgwODNmMDk4MGVjYzI2MTQ4MjhiNzQ1ODcyY2UxOTM3NDI3NThhZDVlNzIwMmQiLCJpYXQiOjE2MDc0NDYxODUsIm5iZiI6MTYwNzQ0NjE4NSwiZXhwIjoxNjM4OTgyMTg1LCJzdWIiOiIyIiwic2NvcGVzIjpbXX0.fcnYNE8v8KQX_K_1WZZB6MWWTS3IASu1CV8RmyBXKc4i0aAw_3OfU_91InzLw3PJrPRJm-pmSPFGb19E9J_QyOGOPBwrQoaUIF6Ir3PTFBEpT8P9geswFz3w53HQcqF954MgwM9-JI7aZYsqHRdJu1blUd8vquuhPsLVT0EKvKT7kIZir1IutuArQq1-xTtb_myh7I_r0-s1GhFEh-wBYU7R5GHFyAT_qmYDns9M_J-suLWMeXcugBDiYOBTrWFv_54TwJIW8L-RYrRXdetekn0ndxXqn4StPvAUpXTBnpqTphPHF2o8lIy0wptCgzrQsV5VXzkj80pbWPZDY01O-I1aMpHSCDwzU4BH3It2LLxLezsln881NkPRD5d-hXQeDCJ88ZnTGk5H3pcUxGrhHs8dYFU9CQ9C1EeKuVYt3XsT1Cbo9Ha-qN-xly_4WbIbNnltnE_AOo92PvWeOXrvSoFCbi6yTKR_ygJskFDW5menTIzEbKbmcvwhfdDPU6iN_V7nA2uQXGt6ApKzj5WsXwUiTaZpPkivEc7n1dF_nriz93AsaVtrB_dmuEsS28q8zIMt3F9uYSwwFS0jL3vWvGhxUBmWy440I6RePWEnTc_fTW-o_BC-iBvhNspwGyR4QySAnP48aMPD0muHij_Q-ZEMafKR_26kFSUC08UXFOA",
                                        "token_type": "Bearer"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}